Create login-view.jsx
Add a login view to your myFlix API, allowing your users to log in to the app. Refer to your API and business logic to make sure you’re asking your users for the right information in the login form. Apply form validation and authentication logic to your login form. Ensure the login details are stored in localStorage so the login session persists until the user logs out.