Update main-view.jsx
Add a logout button to your main view so logged in users can log out. Integrate the login and signup view with MainView so they’re displayed as the first screen for non-authenticated users. Upon login, users should see both the list of movies and the button to log out. Make sure to send the authentication token along with the fetch requests in main-view.jsx to get all movies from your authenticated /movies endpoint on the backend.